refactor: handler 패키지 이름을 core로 변경
